refactor: change facing fields from u8 to Direction enum
Actor.facing, Npc.facing, and GameState.facing are now typed as
Direction instead of raw u8. All call sites updated to use Direction
variants or cast with 'as u8' where arithmetic or u8 interfaces
require it.

Files changed:
- actor.rs: facing: u8 → Direction
- npc.rs: facing: u8 → Direction; Npc::from_bytes init; tick_with_actors casts
- game_state.rs: facing: u8 → Direction; GameState::new() init
- npc_ai.rs: assignments via Direction::from(); Dark Knight facing; tests
- encounter.rs: spawn facing inits; added Direction import
- magic.rs: test init; moved Direction import into #[cfg(test)]
- persist.rs: Direction::from() on load
- debug_tui/bridge.rs: as u8 casts for ActorSnapshot fields
- main.rs: hero_facing as u8 cast
- hiscreen.rs: facing as u8 cast
- gameplay_scene/actors.rs: dragon facing Direction::SE
- gameplay_scene/carriers.rs: Direction::from() on facing_toward result; as u8 probe cast
- gameplay_scene/combat_logic.rs: as u8 cast for Combatant.facing
- gameplay_scene/debug_commands.rs: Direction::NW init
- gameplay_scene/input.rs: tuple type annotation; final_facing type
- gameplay_scene/menu_actions.rs: as u8 cast for fire_missile
- gameplay_scene/rendering.rs: as u8 casts for facing_to_frame_base calls
- gameplay_scene/scene_impl.rs: as u8 casts; Direction variants in matches!
- gameplay_scene/tests.rs: Direction variants throughout; format {:?}

All 733 tests pass.

+++ src/game/hiscreen.rs
@@ -26,7 +26,7 @@ pub fn format_hiscreen(state: &GameState) -> String {
         "VIT:{:3} GOLD:{:4} DIR:{} REGION:{:2} LIGHT:{:3}",
         state.vitality,
         state.gold,
-        facing_char(state.facing),
+        facing_char(state.facing as u8),
         state.region_num,
         state.lightlevel,
     )
